requirements.txt
openai>=1.3.0,<2.0
python-dotenv>=1.0.0,<2.0
